requirements.txt
antiorm==1.2.1
anyio==3.6.2
app==0.0.1
astroid==2.12.13
black==22.10.0
camelcase==0.2
Cerberus==1.3.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
db==0.1.1
dill==0.3.6
dnspython==2.2.1
docopt==0.6.2
fastapi==0.88.0
Flask==2.2.2
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
pathspec==0.10.2
pipreqs==0.4.11
platformdirs==2.6.0
protobuf==3.20.1
psycopg2-binary==2.9.5
pydantic==1.10.2
pylint==2.15.8
pymongo==4.3.2
python-dotenv==0.21.0
python-slugify==7.0.0
requests==2.28.1
sniffio==1.3.0
starlette==0.22.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
Werkzeug==2.2.2
wrapt==1.14.1
yarg==0.1.9
zipp==3.10.0

psycopg2~=2.9.5
SQLAlchemy~=1.4.45
httpx~=0.23.1